Ajustando algumas coisas

+++ API/DTOs/FeedBackUserDto.cs
@@ -7,7 +7,7 @@ namespace API.DTOs
 {
     public class FeedBackUserDto
     {
-        public int Star { get; set; }
+        public double Star { get; set; }
         public string? FeedBack { get; set; }
         public string? UserName { get; set; }
         public DateTime Created { get; set; } = DateTime.UtcNow;
